Añadido info a Degree.
