+++ src/main/java/com/example/tfg/domain/info/CompetenceInfo.java
@@ -1,0 +1,5 @@
+package com.example.tfg.domain.info;
+
+public class CompetenceInfo {
+
+}
