+++ src/main/java/com/example/tfg/web/formatter/DegreeFormatter.java
@@ -21,7 +21,7 @@ public class DegreeFormatter implements Formatter<Degree> {
 	// fetching from Database
 
 	public String print(Degree degree, Locale arg1) {
-		return degree.getName();
+		return degree.getInfoDegree().getName();
 	}
 
 	public Degree parse(String degreeId, Locale arg1) throws ParseException {
