+++ src/main/resources/jdbc.properties
@@ -1,8 +1,8 @@
 jdbc.driverClassName=com.mysql.jdbc.Driver
 
 
-jdbc.url=jdbc:mysql://192.168.205.140:3306/tfg1415
-#jdbc.url=jdbc:mysql://192.168.189.139:3306/tfg1415
+#jdbc.url=jdbc:mysql://192.168.205.140:3306/tfg1415
+jdbc.url=jdbc:mysql://192.168.189.139:3306/tfg1415
 jdbc.username=badgemaker
 jdbc.password=secret
 hibernate.hbm2ddl.auto=create-drop
