+++ src/main/webapp/WEB-INF/views/degree/modify.jsp
@@ -24,16 +24,16 @@
 			<form:form method="post" commandName="modifyDegree" role="form">
 				<div class="form-group">
 					<label>Code: </label>
-					<form:input path="code" placeholder="Code of the subject" class="form-control" required="true"/>
+					<form:input path="infoDegree.code" placeholder="Code of the subject" class="form-control" required="true"/>
 				</div>
 				<div class="form-group">
 					<label>Name: </label>
-					<form:input path="name" class="form-control"
+					<form:input path="infoDegree.name" class="form-control"
 						placeholder="Name of the degree" required="true"/>
 				</div>
 				<div class="form-group">
 					<label>Description: </label>
-					<form:input path="description" class="form-control"
+					<form:input path="infoDegree.description" class="form-control"
 						placeholder="Description of the degree" required="true"/>
 				</div>
